Refactor and enhance type annotations across multiple modules
- Updated type annotations for variables and function return types in Python extractors and compilers to improve type safety and clarity.
- Changed the way parent maps are set in AST trees for better attribute handling.
- Ensured consistent handling of file paths as strings in various repository and operational modules.
- Added assertions to validate the presence of expected attributes and types in context and repository handling.
- Refined the handling of optional values and defaults in various data structures to prevent potential runtime errors.
- Improved the clarity of code by using explicit type casting where necessary, particularly in repository query and indexing operations.

+++ benchmark/run_phase2_profiling.py
@@ -36,7 +36,7 @@ class DiagnosticProfiler:
 
     def get_rss(self) -> float:
         gc.collect()
-        return self.process.memory_info().rss / (1024 * 1024)
+        return float(self.process.memory_info().rss / (1024 * 1024))
 
     def count_reachable_source(self) -> tuple[int, int]:
         import gc
